+++ tests/test030.py
@@ -62,8 +62,8 @@ m3c2_params_dic["UseOriginalCloud"] = "false"
 m3c2_params_dic["UsePrecisionMaps"] = "false"
 m3c2_params_dic["UseSinglePass4Depth"] = "false"
 
-paramFilename = os.path.join(dataDir, "m3c2_params.txt")
-with open(paramFilename, 'w') as f:
+paramFilename1 = os.path.join(dataDir, "m3c2_params1.txt")
+with open(paramFilename1, 'w') as f:
     f.write("[General]\n")
     for k,v in m3c2_params_dic.items():
         f.write("%s=%s\n"%(k,v))
@@ -74,6 +74,8 @@ if cc.isPluginM3C2():
     import cloudComPy.M3C2
     cloud = cc.loadPointCloud(getSampleCloud(5.0))
     cloud1 = cc.loadPointCloud(getSampleCloud(1.0))
+    paramFilename = os.path.join(dataDir, "m3c2_params.txt")
+    ret = cc.M3C2.M3C2gessParamsToFile([cloud,cloud1], paramFilename, True)
     cloud2 = cc.M3C2.computeM3C2([cloud,cloud1], paramFilename)
 #---computeM3C2_01-end